@@ -21,18 +21,18 @@ let otherFuncVoidToVoid = FuncKey<Void, Void>(funcId: "otherFuncVoidToVoid")
 let otherFuncStringToOptionalInt = FuncKey<String, Int?>(funcId: "otherFuncStringToOptionalInt")
 let otherFuncVoidToOptionalInt = FuncKey<Void, Int?>(funcId: "otherFuncVoidToOptionalInt")
 
-var defaultFuncStringToIntGetCall = false
-var defaultFuncStringToVoidGetCall = false
-var defaultFuncVoidToIntGetCall = false
-var defaultFuncVoidToVoidGetCall = false
-var defaultFuncStringToOptionalIntGetCall = false
-var defaultFuncVoidToOptionalIntGetCall = false
-var otherDefaultFuncStringToIntGetCall = false
-var otherDefaultFuncStringToVoidGetCall = false
-var otherDefaultFuncVoidToIntGetCall = false
-var otherDefaultFuncVoidToVoidGetCall = false
-var otherDefaultFuncStringToOptionalIntGetCall = false
-var otherDefaultFuncVoidToOptionalIntGetCall = false
+nonisolated(unsafe) var defaultFuncStringToIntGetCall = false
+nonisolated(unsafe) var defaultFuncStringToVoidGetCall = false
+nonisolated(unsafe) var defaultFuncVoidToIntGetCall = false
+nonisolated(unsafe) var defaultFuncVoidToVoidGetCall = false
+nonisolated(unsafe) var defaultFuncStringToOptionalIntGetCall = false
+nonisolated(unsafe) var defaultFuncVoidToOptionalIntGetCall = false
+nonisolated(unsafe) var otherDefaultFuncStringToIntGetCall = false
+nonisolated(unsafe) var otherDefaultFuncStringToVoidGetCall = false
+nonisolated(unsafe) var otherDefaultFuncVoidToIntGetCall = false
+nonisolated(unsafe) var otherDefaultFuncVoidToVoidGetCall = false
+nonisolated(unsafe) var otherDefaultFuncStringToOptionalIntGetCall = false
+nonisolated(unsafe) var otherDefaultFuncVoidToOptionalIntGetCall = false
 
 let defaultFuncStringToInt = DefaultFuncKey<String, Int>(funcId: "funcStringToInt") { _ in
     defaultFuncStringToIntGetCall = true
@@ -84,29 +84,29 @@ enum Funcs {
         Int(str)!
     }
     
-    static var stringToVoidGetCall = false
+    nonisolated(unsafe) static var stringToVoidGetCall = false
     static func stringToVoid(_ str: String) {
         stringToVoidGetCall = true
     }
     
-    static var voidToIntValue = 1
+    nonisolated(unsafe) static var voidToIntValue = 1
     static func voidToInt() -> Int {
         voidToIntValue
     }
     
-    static var voidToVoidGetCall = false
+    nonisolated(unsafe) static var voidToVoidGetCall = false
     static func voidToVoid() {
         voidToVoidGetCall = true
     }
     
-    static var stringToOptionalIntGetCall = false
+    nonisolated(unsafe) static var stringToOptionalIntGetCall = false
     static func stringToOptionalInt(_ str: String) -> Int? {
         stringToOptionalIntGetCall = true
         return Int(str)
     }
     
-    static var voidToOptionalIntGetCall = false
-    static var voidToOptionalIntValue: Int? = nil
+    nonisolated(unsafe) static var voidToOptionalIntGetCall = false
+    nonisolated(unsafe) static var voidToOptionalIntValue: Int? = nil
     static func voidToOptionalInt() -> Int? {
         voidToOptionalIntGetCall = true
         return voidToOptionalIntValue
